pyproject.toml
[project]
name = "Ai-Agent Powered EDA"
version = "0.1.0"
description = "Data Analytics Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "groq>=0.32.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.10",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.0",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
